+++ app/frontend/containers/NotificationMenuContainer.js
@@ -1,0 +1,47 @@
+import React from 'react';
+import { connect } from 'react-redux';
+import NotificationMenu from '../ui/NotificationMenu';
+import ApiClient from './ApiClient';
+import fetchNotifications from '../actions/fetchNotifications';
+import readNotification from '../actions/readNotification';
+import sleep from '../utils/sleep';
+
+const mapStateToProps = state => {
+  return {
+    currentUser: state.app.currentUser,
+    large: state.shared.large,
+    notifications: state.shared.notifications,
+    unreadNotifications: state.shared.unreadNotifications
+  };
+};
+
+const mapDispatchToProps = (dispatch, ownProps) => {
+  return {
+    handleMount: async () => {
+      const client = new ApiClient();
+      let response = await client.fetchNotifications();
+      if (response.ok) {
+        let notifications = await response.json();
+        dispatch(fetchNotifications(notifications));
+      }
+    },
+
+    readNotifications: async notifications => {
+      await sleep(5000);
+      const client = new ApiClient();
+      let unreadNotifications = notifications.filter(notification => {
+        return notification.read === false;
+      });
+      unreadNotifications.forEach(async notification => {
+        let response = await client.readNotification(notification.id);
+        if (response.ok) {
+          let notification = await response.json();
+          dispatch(readNotification(notification));
+        }
+        await sleep(3000);
+      });
+    }
+  };
+};
+
+export default React.memo(connect(mapStateToProps, mapDispatchToProps)(NotificationMenu));
